% Generated by roxygen2 (4.0.0): do not edit by hand
\name{anyInf}
\alias{anyInf}
\title{Identify if any infinite values are present in a vector}
\usage{
anyInf(x)
}
\arguments{
\item{x}{a vector of any class that contains Inf values}
}
\value{
a logical, TRUE if infinite values are present, and FALSE if not
}
\description{
A finite analog of \code{\link{anyNA}} that returns a logical
}
\note{
Yadda yadda yadda
}
\author{
Jared E. Knowles
}

