% Generated by roxygen2 (4.0.0): do not edit by hand
\name{MAR.df}
\alias{MAR.df}
\title{Assign NAs to columns in a dataframe at random}
\usage{
MAR.df(df, vars, probs)
}
\arguments{
\item{df}{a dataframe that missing data should be supplied to}

\item{probs}{A single value from 0 to 1 representing the proportion of each column
that should be replaced with NAs, or a list of such values length of \code{vars}}

\item{vars}{A list of variable names to be used in generating the probability models
for the missing data}
}
\value{
df with proportion p of values replaced with NA
}
\description{
Create a dataframe with data missing at random by generating probability models
of missing data from observable characteristics and then eliminating data based
on those results.
}
\note{
Yadda yadda yadda
}
\author{
Jared E. Knowles
}

