% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingness.R
\name{MCARplot}
\alias{MCARplot}
\title{MCAR Plot}
\usage{
MCARplot(results)
}
\arguments{
\item{results}{results from MCARcheck.df}
}
\value{
ggplot2 object
}
\description{
MCAR Plot
}
\author{
Jared E. Knowles
}

