% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingness.R
\name{MCARx}
\alias{MCARx}
\title{Assign NAs to a vector randomly}
\usage{
MCARx(x, p)
}
\arguments{
\item{x}{a vector of any class that NA values should be supplied to}

\item{p}{A value from 0 to 1 representing the proportion of x that should be replaced with NAs}
}
\value{
x with proportion p of values replaced with NA
}
\description{
Assign NAs to a vector randomly based on a user specified probability
}
\note{
Yadda yadda yadda
}
\author{
Jared E. Knowles
}

