% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rnormcor}
\alias{rnormcor}
\title{Generate a single variable from a distribution correlated with another distribution}
\usage{
rnormcor(x, rho)
}
\arguments{
\item{x}{variable to draw from}

\item{rho}{correlation coefficient between x and result of function}
}
\value{
a single numeric value drawn from a normal distribution correlated with x at the level of rho
}
\description{
Allow user to draw from a random normal distribution correlated with a user specified distribution
}
\details{
Rough estimate
}
\author{
Jared E. Knowles
}

