\name{summary.MCARcheck}
\alias{summary.MCARcheck}
\title{Summary for MCAR checks}
\usage{
  summary.MCARcheck(results, p, print)
}
\arguments{
  \item{results}{results from MCARcheck.df}

  \item{p}{threshold for statistical significance on the
  gamma}
}
\value{
  summary of dependencies among missing values in dataset
}
\description{
  Summary for MCAR checks
}
\author{
  Jared E. Knowles
}

