% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingness.R
\name{vecInfSearch}
\alias{vecInfSearch}
\title{Count the proportion of Inf values in a vector}
\usage{
vecInfSearch(x)
}
\arguments{
\item{x}{a vector of any class that contains Inf values}
}
\value{
the proportion of x that are Inf values
}
\description{
Return the proportion of vector x that are Inf
}
\note{
Yadda yadda yadda
}
\author{
Jared E. Knowles
}

