% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rbinomcor}
\alias{rbinomcor}
\title{Generate a vector from a binomial distribution correlated with another distribution}
\usage{
rbinomcor(x, rho, scale = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{variable to draw from}

\item{rho}{correlation coefficient between x and result of function}

\item{scale}{a scale factor for the binomial draws}

\item{na.rm}{a logical indicating whether to fit the distribution excluding missing values
or to fail on missing values}
}
\value{
a vector of the same length as x drawn from a binomial distribution correlated with x at the level of rho
}
\description{
Allow user to draw from a binomial distribution correlated with a user specified vector
}
\details{
Rough estimate
}
\author{
Jared E. Knowles
}

