% Generated by roxygen2 (4.0.0): do not edit by hand
\name{summary.MCARcheck}
\alias{summary.MCARcheck}
\title{Summary for MCAR checks}
\usage{
\method{summary}{MCARcheck}(results, p, print)
}
\arguments{
\item{results}{results from MCARcheck.df}

\item{p}{threshold for statistical significance on the gamma}

\item{print}{Option to print out results}
}
\value{
summary of dependencies among missing values in dataset
}
\description{
Summary for MCAR checks
}
\author{
Jared E. Knowles
}

